+++ src/quip_entity/tree_node.py
